Parse in full post data, with comments
Updated the parser so it will load in the full XML, create a new array for all the
posts, then parse each post in the XML into that array and create objects out of it.
This lets you do a few nice things such as use the .reject and .select Enumerable
methods to filter out any posts you don't want to keep or migrate.